Skypro_PD_13.0_Ilyasov/requirements.txt
APScheduler==3.10.1
Django==4.1.7
django-allauth==0.52.0
django-apscheduler==0.6.2
django-cors-headers==3.14.0
django-filter==22.1
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.1.0
drf-nested-routers==0.93.4
drf-spectacular==0.26.0
Pillow==9.4.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0